Moves Notification to a separate file
