+++ apps/judicial-system/api/src/app/modules/case/dto/sendNotification.input.ts
@@ -2,7 +2,7 @@ import { Allow } from 'class-validator'
 
 import { Field, InputType } from '@nestjs/graphql'
 
-import {
+import type {
   NotificationType,
   SendNotification,
 } from '@island.is/judicial-system/types'
