+++ apps/judicial-system/api/src/app/modules/case/models/sendNotification.response.ts
@@ -1,6 +1,6 @@
 import { Field, ObjectType } from '@nestjs/graphql'
 
-import { SendNotificationResponse as TSendNotificationResponse } from '@island.is/judicial-system/types'
+import type { SendNotificationResponse as TSendNotificationResponse } from '@island.is/judicial-system/types'
 
 @ObjectType()
 export class SendNotificationResponse implements TSendNotificationResponse {
