+++ apps/judicial-system/backend/src/app/modules/notification/models/sendNotification.resopnse.ts
@@ -1,6 +1,6 @@
 import { ApiProperty } from '@nestjs/swagger'
 
-import { SendNotificationResponse as TSendNotificationResponse } from '@island.is/judicial-system/types'
+import type { SendNotificationResponse as TSendNotificationResponse } from '@island.is/judicial-system/types'
 
 import { Notification } from './notification.model'
