+++ libs/judicial-system/types/src/index.ts
@@ -3,18 +3,22 @@ export { Feature } from './lib/feature'
 export { InstitutionType } from './lib/institution'
 export type { Institution } from './lib/institution'
 
+export { NotificationType } from './lib/notification'
+export type {
+  Notification,
+  SendNotification,
+  SendNotificationResponse,
+} from './lib/notification'
+
 export { UserRole } from './lib/user'
 export type { User, CreateUser, UpdateUser } from './lib/user'
 
 export * from './lib/types'
 export type {
   Case,
-  Notification,
   CreateCase,
   UpdateCase,
   TransitionCase,
-  SendNotification,
-  SendNotificationResponse,
   RequestSignatureResponse,
   SignatureConfirmationResponse,
   CreateCourtCase,
